pyproject.toml
[tool.poetry]
name = "qtchair-pyside"
version = "0.1.0"
description = ""
license = "AGPL-3.0"
authors = ["cloudw233 <qrz20081004@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pyside6-fluent-widgets = {extras = ["full"], version = "^1.7.5"}
azure-cognitiveservices-speech = "^1.42.0"
qasync = "^0.27.1"
apscheduler = "^3.11.0"
websockets = "^15.0.1"
orjson = "^3.10.16"
tomlkit = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
